requirements.txt
Flask==2.2.5
flask-cors==3.0.10
requests==2.31.0
numpy==1.26.0
Pillow==10.0.0
rasterio==1.3.9
python-dotenv
psycopg2-binary        # or async driver if using SQLAlchemy async
SQLAlchemy
geoalchemy2
redis
rq                    # optional: Redis Queue
firebase-admin        # optional: if verifying Firebase tokens
alembic               # migrations
gunicorn
shapely
botocore
boto3

